change Period type in api response to long

+++ CryptoApp.Infrastructure/API/CoinCap/Responses/CandlePoint.cs
@@ -6,5 +6,5 @@ public record CandlePoint(
     decimal Low,
     decimal Close,
     decimal Volume,
-    decimal Period
+    long Period
     );
